Separate Narrator Service from Notification creation logic.
